[feat/#71] 내 시야 아카이빙 임시 적용 (commit body TODO list)
해결해야할 TODO 리스트
- 게시물 수정, 삭제시 위치 초기화 해결
- mutliviewType recyclerview에서 stickyHeader 적용하기
- 로딩, 실패시 적용하기
