+++ presentation/src/main/java/com/depromeet/presentation/seatrecord/SeatRecordActivity.kt
@@ -3,39 +3,24 @@ package com.depromeet.presentation.seatrecord
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import android.text.SpannableString
-import android.text.Spanned
-import android.text.style.ForegroundColorSpan
-import android.view.View
-import android.view.View.GONE
-import android.view.View.VISIBLE
-import android.widget.AdapterView
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.fragment.app.commit
 import androidx.lifecycle.asLiveData
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import coil.load
-import coil.transform.CircleCropTransformation
+import androidx.recyclerview.widget.RecyclerView.OnScrollListener
 import com.depromeet.core.base.BaseActivity
 import com.depromeet.core.state.UiState
-import com.depromeet.designsystem.SpotSpinner
-import com.depromeet.designsystem.extension.dpToPx
 import com.depromeet.domain.entity.response.home.MySeatRecordResponse
-import com.depromeet.domain.entity.response.home.ReviewDateResponse
 import com.depromeet.presentation.R
 import com.depromeet.presentation.databinding.ActivitySeatRecordBinding
 import com.depromeet.presentation.extension.toast
 import com.depromeet.presentation.home.ProfileEditActivity
 import com.depromeet.presentation.seatReview.ReviewActivity
-import com.depromeet.presentation.seatrecord.adapter.DateMonthAdapter
-import com.depromeet.presentation.seatrecord.adapter.LinearSpacingItemDecoration
-import com.depromeet.presentation.seatrecord.adapter.MonthRecordAdapter
-import com.depromeet.presentation.seatrecord.uiMapper.MonthReviewData
+import com.depromeet.presentation.seatrecord.test.RecordListItem
+import com.depromeet.presentation.seatrecord.test.SeatRecordAdapter
 import com.depromeet.presentation.seatrecord.viewmodel.EditUi
 import com.depromeet.presentation.seatrecord.viewmodel.SeatRecordViewModel
-import com.depromeet.presentation.util.CalendarUtil
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
 
@@ -45,18 +30,15 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
 ) {
     companion object {
         const val SEAT_RECORD_TAG = "seatRecord"
-        private const val START_SPACING_DP = 20
-        private const val BETWEEN_SPACING_DP = 8
         const val PROFILE_NAME = "profile_name"
         const val PROFILE_IMAGE = "profile_image"
         const val PROFILE_CHEER_TEAM = "profile_cheer_team"
     }
 
-    private lateinit var dateMonthAdapter: DateMonthAdapter
-    private lateinit var monthRecordAdapter: MonthRecordAdapter
+    private lateinit var adapter: SeatRecordAdapter
     private val viewModel: SeatRecordViewModel by viewModels()
     private var isLoading: Boolean = false
-    private var isSpinnerInitialized = false
+
 
     private val editProfileLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
@@ -80,10 +62,8 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
     }
 
     private fun initView() {
-        initMonthAdapter()
-        setRefreshClicked()
-        initReviewList()
         viewModel.getReviewDate()
+        initRecordAdapter()
     }
 
     private fun initEvent() {
@@ -92,24 +72,13 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
                 finish()
             }
             fabRecordUp.setOnClickListener {
-                ssvRecord.smoothScrollTo(0, 0)
+                rvSeatRecord.smoothScrollToPosition(0)
             }
             fabRecordPlus.setOnClickListener {
                 Intent(this@SeatRecordActivity, ReviewActivity::class.java).apply {
                     startActivity(
                         this
                     )
-                }
-            }
-            btRecordFailResfresh.setOnClickListener {
-                viewModel.getSeatRecords()
-            }
-            ivRecordEdit.setOnClickListener {
-                navigateToProfileEditActivity()
-            }
-            btRecordWriteRecord.setOnClickListener {
-                Intent(this@SeatRecordActivity, ReviewActivity::class.java).apply {
-                    startActivity(this)
                 }
             }
         }
@@ -121,13 +90,50 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         observeEvents()
     }
 
-    private fun setRefreshClicked() {
-        binding.btRecordFailResfresh.setOnClickListener {
-            if (viewModel.date.value is UiState.Success) {
-                viewModel.getSeatRecords()
-            } else {
-                viewModel.getReviewDate()
+    private fun initRecordAdapter() {
+        adapter = SeatRecordAdapter(
+            reviewClick = {
+                viewModel.setClickedReviewId(it.id)
+                supportFragmentManager.commit {
+                    replace(
+                        R.id.fcv_record,
+                        SeatDetailRecordFragment(),
+                        SeatDetailRecordFragment.SEAT_RECORD_TAG
+                    )
+                    addToBackStack(null)
+                }
+            },
+            monthClick = { month ->
+                viewModel.setSelectedMonth(month)
+                binding.rvSeatRecord.smoothScrollToPosition(0)
+            },
+            yearClick = { year ->
+                viewModel.setSelectedYear(year)
+                binding.rvSeatRecord.smoothScrollToPosition(0)
+            },
+            profileEditClick = {
+                navigateToProfileEditActivity()
+            },
+            reviewEditClick = { reviewId ->
+                viewModel.setEditReviewId(reviewId)
+                RecordEditDialog.newInstance(SEAT_RECORD_TAG)
+                    .apply { show(supportFragmentManager, this.tag) }
             }
+        )
+
+        with(binding) {
+            rvSeatRecord.adapter = adapter
+            rvSeatRecord.addOnScrollListener(object : OnScrollListener() {
+                override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
+                    val scrollBottom = !binding.rvSeatRecord.canScrollVertically(1)
+                    if (scrollBottom && !isLoading && viewModel.reviews.value is UiState.Success) {
+                        if (!(viewModel.reviews.value as UiState.Success).data.last) {
+                            isLoading = true
+                            viewModel.loadNextSeatRecords()
+                        }
+                    }
+                }
+            })
         }
     }
 
@@ -135,25 +141,33 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         viewModel.date.asLiveData().observe(this) { state ->
             when (state) {
                 is UiState.Success -> {
-                    setReviewsVisibility(isExist = true)
-                    setYearSpinner(state.data)
-                    dateMonthAdapter.submitList(state.data.yearMonths.first { it.isClicked }.months)
+                    val reviewState = viewModel.reviews.value
+                    if (reviewState is UiState.Success) {
+                        adapter.updateItemAt(1, RecordListItem.Date(state.data.yearMonths))
+                    } else {
+                        adapter.submitList(
+                            listOf(
+                                RecordListItem.Profile(MySeatRecordResponse.MyProfileResponse()),
+                                RecordListItem.Date(state.data.yearMonths),
+                                RecordListItem.Record(emptyList())
+                            )
+                        )
+                    }
                     viewModel.getSeatRecords()
                 }
 
                 is UiState.Empty -> {
-                    setReviewsVisibility(isExist = false)
-                    binding.shimmerRecord.visibility = GONE
+                    //TODO : 기록 SHIMMER 처리 GONE
+                    Timber.d("test empty")
                 }
 
                 is UiState.Loading -> {
-                    setReviewsVisibility(isExist = true)
+                    Timber.d("test loading")
                 }
 
                 is UiState.Failure -> {
-                    setReviewsVisibility(isExist = true)
-                    binding.clHomeFail.visibility = VISIBLE
-                    binding.rvRecordMonthDetail.visibility = GONE
+                    //TODO : 기록 SHIMMER 처리 GONE + 실패 LAYOUT 보여줘야함
+                    Timber.d("test fail")
                 }
 
             }
@@ -164,27 +178,31 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
         viewModel.reviews.asLiveData().observe(this) { state ->
             when (state) {
                 is UiState.Success -> {
-                    setProfile(state.data.profile)
-                    updateReviewList(state.data.reviews)
-                    binding.clHomeFail.visibility = GONE
-                    binding.rvRecordMonthDetail.visibility = VISIBLE
+
+                    Timber.d("test왜 다른거지? \n ${state.data.reviews}\n ${state.data.profile}")
+                    val newProfileItem = RecordListItem.Profile(state.data.profile)
+                    val newRecordItem = RecordListItem.Record(state.data.reviews)
+                    val newList = adapter.currentList.toMutableList()
+                    newList[0] = newProfileItem
+                    newList[2] = newRecordItem
+
+
+                    adapter.submitList(newList)
+
                     isLoading = false
-                    setShimmer(false)
+                    //TODO : 실패 VISIBILITY GONE + 리사이클러뷰 보여줘야함
                 }
 
                 is UiState.Loading -> {
-                    setShimmer(true)
                 }
 
                 is UiState.Empty -> {
-                    monthRecordAdapter.submitList(emptyList())
+                    //TODO monthadapter에 비어있는 리스트 보내줘야함
                     toast("해당 월에 작성한 글이 없습니다.")
                 }
 
                 is UiState.Failure -> {
-                    setShimmer(false)
-                    binding.clHomeFail.visibility = VISIBLE
-                    binding.rvRecordMonthDetail.visibility = GONE
+                    //TODO 실패 보여줘야하고 리사이클러뷰 사라지게?
                 }
             }
         }
@@ -202,176 +220,6 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
                 moveEditReview()
             }
         }
-    }
-
-    private fun setProfile(data: MySeatRecordResponse.MyProfileResponse) {
-        with(binding) {
-            if (data.teamId != null) {
-                "${data.teamName}의 Lv.${data.level} ${data.levelTitle}".also {
-                    csbvRecordTitle.setText(
-                        it
-                    )
-                }
-            } else {
-                "모두를 응원하는 Lv.${data.level} ${data.levelTitle}".also { csbvRecordTitle.setText(it) }
-            }
-            tvRecordNickname.text = data.nickname
-            tvRecordCount.text = data.reviewCount.toString()
-            ivRecordProfile.load(data.profileImage) {
-                transformations(CircleCropTransformation())
-                error(com.depromeet.designsystem.R.drawable.ic_default_profile)
-            }
-        }
-    }
-
-    private fun createColoredLevelString(data: MySeatRecordResponse.MyProfileResponse): SpannableString {
-        /** 커스텀뷰에 또 spannable을 적용해야한다니... */
-        val fullText = if (data.teamId != null) {
-            "${data.teamName}의 Lv.${data.level} ${data.levelTitle}"
-        } else {
-            "모두를 응원하는 Lv.${data.level} ${data.levelTitle}"
-        }
-        val spannableString = SpannableString(fullText)
-
-        val levelStartIndex = fullText.indexOf(data.level.toString())
-        val levelEndIndex = levelStartIndex + data.level.toString().length
-
-        spannableString.setSpan(
-            ForegroundColorSpan(getColor(com.depromeet.designsystem.R.color.color_action_enabled)),
-            levelStartIndex,
-            levelEndIndex,
-            Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-        )
-
-        return spannableString
-    }
-
-    private fun setYearSpinner(data: ReviewDateResponse) {
-        if (isSpinnerInitialized) return
-        val years = data.yearMonths.map { it.year }
-        val yearList = years.map { "${it}년" }
-
-        val adapter = SpotSpinner(
-            this,
-            com.depromeet.designsystem.R.layout.item_spot_spinner_view,
-            com.depromeet.designsystem.R.layout.item_spot_spinner_dropdown,
-            yearList,
-            true,
-            com.depromeet.designsystem.R.color.color_foreground_heading,
-        )
-        with(binding.spinnerRecordYear) {
-            this.adapter = adapter
-            onItemSelectedListener =
-                object : AdapterView.OnItemSelectedListener {
-                    override fun onItemSelected(
-                        parent: AdapterView<*>?,
-                        view: View?,
-                        position: Int,
-                        id: Long,
-                    ) {
-                        adapter.setSelectedItemPosition(position)
-                        val selectedYear = yearList[position].filter { it.isDigit() }.toInt()
-                        viewModel.setSelectedYear(selectedYear)
-                        binding.ssvRecord.smoothScrollTo(0, 0)
-                    }
-
-                    override fun onNothingSelected(p0: AdapterView<*>?) {}
-                }
-        }
-        isSpinnerInitialized = true
-    }
-
-    private fun initMonthAdapter() {
-        dateMonthAdapter = DateMonthAdapter(
-            monthClick = { month ->
-                viewModel.setSelectedMonth(month)
-                binding.ssvRecord.smoothScrollTo(0, 0)
-            }
-        )
-        binding.rvRecordMonth.adapter = dateMonthAdapter
-        binding.rvRecordMonth.addItemDecoration(
-            LinearSpacingItemDecoration(
-                START_SPACING_DP.dpToPx(this),
-                BETWEEN_SPACING_DP.dpToPx(this)
-            )
-        )
-    }
-
-    private fun setReviewsVisibility(isExist: Boolean) {
-        if (!isExist) {
-            with(binding) {
-                "${CalendarUtil.getCurrentYear()}년".also { tvRecordYear.text = it }
-                clRecordNone.visibility = VISIBLE
-                clRecordStickyHeader.visibility = GONE
-                rvRecordMonthDetail.visibility = GONE
-            }
-        } else {
-            with(binding) {
-                clRecordNone.visibility = GONE
-                clRecordStickyHeader.visibility = VISIBLE
-                rvRecordMonthDetail.visibility = VISIBLE
-
-                ssvRecord.header = binding.clRecordStickyHeader
-            }
-        }
-    }
-
-    private fun initReviewList() {
-        monthRecordAdapter = MonthRecordAdapter()
-        binding.rvRecordMonthDetail.adapter = monthRecordAdapter
-        binding.rvRecordMonthDetail.itemAnimator = null
-
-        monthRecordAdapter.itemRecordClickListener =
-            object : MonthRecordAdapter.OnItemRecordClickListener {
-
-                override fun onItemRecordClick(item: MySeatRecordResponse.ReviewResponse) {
-                    viewModel.setClickedReviewId(item.id)
-                    supportFragmentManager.commit {
-                        replace(
-                            R.id.fcv_record,
-                            SeatDetailRecordFragment(),
-                            SeatDetailRecordFragment.SEAT_RECORD_TAG
-                        )
-                        addToBackStack(null)
-                    }
-                }
-
-                override fun onMoreRecordClick(reviewId: Int) {
-                    viewModel.setEditReviewId(reviewId)
-                    RecordEditDialog.newInstance(SEAT_RECORD_TAG)
-                        .apply { show(supportFragmentManager, this.tag) }
-                }
-
-            }
-
-        binding.rvRecordMonthDetail.addOnScrollListener(object : RecyclerView.OnScrollListener() {
-            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                super.onScrolled(recyclerView, dx, dy)
-
-                val scrollBottom = !binding.rvRecordMonthDetail.canScrollVertically(1)
-
-                val layoutManager = binding.rvRecordMonthDetail.layoutManager as LinearLayoutManager
-                val lastVisibleItemPosition = layoutManager.findLastVisibleItemPosition()
-                val itemCount = layoutManager.itemCount - 1
-
-                Timber.d("test 마지막 포지션 : $lastVisibleItemPosition 아이템 : $itemCount")
-
-//                val hasNextPage = (viewModel.reviews.value as? UiState.Success)?.data?.last == false
-//                if (lastVisibleItemPosition == itemCount && hasNextPage && !isLoading) {
-//                    isLoading = true
-//                    viewModel.loadNextSeatRecords()
-//                }
-            }
-        })
-    }
-
-    private fun updateReviewList(reviews: List<MySeatRecordResponse.ReviewResponse>) {
-        val groupList =
-            reviews.groupBy { CalendarUtil.getMonthFromDateFormat(it.date) }
-                .map { (month, reviews) ->
-                    MonthReviewData(month, reviews)
-                }
-        monthRecordAdapter.submitList(groupList)
     }
 
     private fun navigateToProfileEditActivity() {
@@ -394,18 +242,7 @@ class SeatRecordActivity : BaseActivity<ActivitySeatRecordBinding>(
 
 
     private fun moveEditReview() {
-
-    }
-
-    private fun setShimmer(isLoading: Boolean) = with(binding) {
-        if (isLoading) {
-            shimmerRecord.startShimmer()
-            shimmerRecord.visibility = VISIBLE
-            shimmerRecordProfile.visibility = VISIBLE
-        } else {
-            shimmerRecord.stopShimmer()
-            shimmerRecord.visibility = GONE
-            shimmerRecordProfile.visibility = GONE
-        }
+        //TODO : 나중에 삭제해야함
+        viewModel.test()
     }
 }
