--- presentation/src/main/res/layout/activity_test_seat_record.xml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/color_background_tertiary">
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/gl_start"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintGuide_begin="16dp" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/gl_end"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintGuide_end="16dp" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/gl_bottom"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintGuide_end="30dp" />
-
-
-    <com.depromeet.designsystem.SpotAppbar
-        android:id="@+id/record_spot_appbar"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:background="@color/color_background_tertiary"
-        android:text="@string/view_record_my"
-        android:textColor="@color/color_foreground_heading"
-        android:textSize="16sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:navigationIcon="@drawable/ic_chevron_left_big"
-        app:navigationIconColor="@color/color_foreground_body_sebtext"
-        app:navigationIconSize="24"
-        app:titleTextAppearance="@style/TextAppearance.Spot.Label03" />
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_seat_record"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:orientation="vertical"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toTopOf="@id/tv_test"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/record_spot_appbar"
-        />
-
-    <TextView
-        android:id="@+id/tv_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="끝에 도달 테스트"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:textSize="40sp"
-        android:textColor="@color/color_foreground_heading"
-        android:visibility="gone"
-        />
-
-    <ImageButton
-        android:id="@+id/fab_record_up"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
-        android:layout_marginBottom="16dp"
-        android:background="@drawable/circle_white_fill"
-        android:elevation="2dp"
-        android:padding="12dp"
-        android:scaleType="center"
-        android:src="@drawable/ic_up"
-        app:layout_constraintBottom_toTopOf="@id/fab_record_plus"
-        app:layout_constraintEnd_toEndOf="@id/gl_end"
-        tools:ignore="contentDescription" />
-
-
-    <ImageButton
-        android:id="@+id/fab_record_plus"
-        android:layout_width="48dp"
-        android:layout_height="48dp"
-        android:background="@drawable/circle_spot_green_fill"
-        android:elevation="2dp"
-        android:padding="12dp"
-        android:scaleType="center"
-        android:src="@drawable/ic_plus_24"
-        app:layout_constraintBottom_toBottomOf="@id/gl_bottom"
-        app:layout_constraintEnd_toEndOf="@id/gl_end"
-        tools:ignore="contentDescription" />
-
-    <FrameLayout
-        android:id="@+id/fcv_record"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:elevation="2dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
